remplacement de vector par list pour colorfig, de de vector par une map dans board. Ajout d'une classe point template pour construire les figures, initializer_list dans la class Point. Modification de shape (couleur différente quand figure selectionné + correction de quelque bug

+++ Rendu-final/include/parallelogram.hpp
@@ -27,7 +27,7 @@ namespace geometricShape{
              *  \param height 
              *  \param width
              */
-            Parallelogram(int sizeCote, int width, int height);
+            Parallelogram(double sizeCote, int width, int height);
 
             /*!
              *  \brief
